+++ client/components/Games/Level0/Game01.js
@@ -12,15 +12,53 @@ export const Game01 = () => {
       {
         kind: 'block',
         type: 'text',
-        fields: { TEXT: 'hello world' },
+        fields: { TEXT: '' },
       },
     ],
   };
 
+  const initApi = (interpreter, scope) => {
+    // Add an API function for the alert() block.
+    const wrapper = function (text) {
+      text = text ? text.toString() : '';
+      const test = document.getElementById('test');
+      test.innerHTML = text;
+
+      // original line below. createPrimitive throwing an error so we adjusted it because 🙄
+      // return interpreter.createPrimitive(alert(text))
+    };
+    interpreter.setProperty(
+      scope,
+      'alert',
+      interpreter.createNativeFunction(wrapper)
+    );
+  };
+
+  //for this ternary we would need to make sure the instructions say to write hello pigeons with no caps or punctuation
+
+  const outcome = () => {
+    document.getElementById('test').innerHTML === 'hello pigeons'
+      ? setTimeout(() => {
+          alert('great job!');
+        }, 500)
+      : alert(
+          'SO CLOSE - try again! HINT: did you make sure to write "hello pigeons" in the block?'
+        );
+  };
+
   return (
-    <>
-      <Workspace toolbox={toolbox} />
-    </>
+    <div style={{ margin: 100 }}>
+      <div
+        style={{
+          height: 100,
+          width: 500,
+          backgroundColor: '#add8e6',
+        }}
+      >
+        <p id="test"></p>
+      </div>
+      <Workspace toolbox={toolbox} initApi={initApi} outcome={outcome} />
+    </div>
   );
 };
