+++ client/components/Games/Workspace.js
@@ -1,11 +1,10 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { BlocklyWorkspace } from 'react-blockly';
 import Blockly from 'blockly';
 import Interpreter from 'js-interpreter';
-// import * as Acorn from 'acorn-node';
 
 const Workspace = (props) => {
-  const { toolbox } = props;
+  const { toolbox, initApi, outcome } = props;
   const [javascriptCode, setJavascriptCode] = useState('');
 
   const workspaceDidChange = (workspace) => {
@@ -14,36 +13,9 @@ const Workspace = (props) => {
   };
 
   const clickHandler = () => {
-    const initApi = (interpreter, scope) => {
-      // Add an API function for the alert() block.
-      const wrapper = function (text) {
-        text = text ? text.toString() : '';
-        return alert(text);
-        // original line below. createPrimitive throwing an error so we adjusted it because 🙄
-        // return interpreter.createPrimitive(alert(text))
-      };
-      interpreter.setProperty(
-        scope,
-        'alert',
-        interpreter.createNativeFunction(wrapper)
-      );
-    };
-
-    // const myInterpreter = new Interpreter(javascriptCode);
-    // const code = javascriptCode.toString();
     const myInterpreter = new Interpreter(javascriptCode, initApi);
-    // console.log('JAVASCRIPT CODE 😤', code);
-    console.log('MY INTERPRETER 😒', myInterpreter);
-    console.log('RUN', myInterpreter.run());
-    // eval(javascriptCode);
     myInterpreter.run();
-    // const nextStep = () => {
-    //   if (myInterpreter.step()) {
-    //     window.setTimeout(nextStep, 0);
-    //   }
-    // };
-    // nextStep();
-    console.log('I CLICKED 🤪');
+    outcome();
   };
 
   //put visuals of game above blockly workspace
